Huge rework of output configuration and set flight mode.
This now uses UBX-CFG "binary" encoded messages.
Contains an implementation of a packet encapsulator which takes a class id, message id and a payload of bytes, encapsulates it with header + length and a checksum at the end, and then looks for the corresponding CFG-ACK from the u-blox chip.
It works!
This sets the first 5 output messages to off, and also enables flight mode using a string I found from Serenity code.
Removed all old PUBX configuration code, this is unncessary now.
Also contains a method within the read thread which differentiates between ublox packets and strings, to read them without looking for a line ending like we get with NMEA sentences.